Got folder and garbage filtering to work properly.
The secret to getting it to work was simply utilizing the correct
properties. Google uses the mimeType field of all things to determine
wether or not something is a file. Additionally, it returns things
that are trashed by default. Does it make sense, no, but it's what we
all have to deal with.